+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
   * Run/Stop option in context menu
+  * Duplicate task option in context menu
   * Fixed delete suite bug
 
 0.5.1 / 2018-06-14